% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublebootstrap.R
\name{db_estimators}
\alias{db_estimators}
\title{Estimators for Double Bootstrap.}
\usage{
db_estimators(data, k)
}
\arguments{
\item{data}{Original data}

\item{k}{Number of top-order statistics}
}
\value{
Returns a list containing values for \eqn{\hat \xi^{(1)}} determined through the Hill estimator, and \eqn{M_{(n,k)}}
}
\description{
Helper function for doublebootstrap(). Calculates values for \eqn{\hat \xi^{(1)}}, and \eqn{M_{(n,k)}}.
}
\details{
\deqn{\hat{\xi}_{n,k}^{(1)} := \frac{1}{k}\sum_{i=1}^{k}\log(\frac{X_{(i)}}{X_{(k+1)}})}

\deqn{M_{n,k} := \frac{1}{k}\sum_{i=1}^{k} (\log(\frac{X_{(i)}}{X_{(k+1)}}))^2}
}
\examples{

x <- heavytails:::generate_expbody_partail(n = 5000, alpha = 2.0)
db_est_values <- heavytails:::db_estimators(data = x, k = 10)

}
\keyword{internal}
