% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doublebootstrap.R
\name{find_db_diffs}
\alias{find_db_diffs}
\title{Difference calculation for Double Bootstrap}
\usage{
find_db_diffs(data, k_seq)
}
\arguments{
\item{data}{Original dataset}

\item{k_seq}{Sequence of top-order statistics}
}
\value{
Returns a vector of squared differences between \eqn{\hat \xi^{(1)}} and \eqn{M_{n,k}}
}
\description{
Calculates the differences between \eqn{\hat \xi^{(1)}} and \eqn{\hat \xi^{(2)}} for each k value obtained through a sequence.
}
\details{
Calculates \eqn{(M_{n,k}(j) - 2(\hat{\xi}_{n,k}^{(1)}(j))^2)^2}
}
\examples{

x <- heavytails:::generate_expbody_partail(n = 5000, alpha = 2.0)
diffs <- heavytails:::find_db_diffs(data = x, k_seq = c(2,100))


}
\keyword{internal}
